Refactor AggroModels to use repository pattern and services
- Extract video operations to VideoRepository
- Extract channel operations to ChannelRepository
- Extract archive logic to ArchiveService
- Extract thumbnail logic to ThumbnailService
- Update phpstan configuration for new classes
- Improve code organization and maintainability
